+++ packages/react/components/input.js
@@ -173,10 +173,16 @@ class F7Input extends React.Component {
     let inputEl;
 
     const createInput = (InputTag, children) => {
-      const needsValue = type !== 'file';
+      const needsValue = type !== 'file' && type !== 'datepicker';
       const needsType = InputTag === 'input';
+      let inputType = type;
+
+      if (inputType === 'datepicker') {
+        inputType = 'text';
+      }
+
       const inputClassName = Utils.classNames(!wrap && className, {
-        resizable: type === 'textarea' && resizable,
+        resizable: inputType === 'textarea' && resizable,
         'no-store-data': noFormStoreData || noStoreData || ignoreStoreData,
         'input-invalid': errorMessage && errorMessageForce || self.state.inputInvalid,
         'input-with-value': inputHasValue,
@@ -190,8 +196,12 @@ class F7Input extends React.Component {
       }
 
       const valueProps = {};
-      if ('value' in props) valueProps.value = inputValue;
-      if ('defaultValue' in props) valueProps.defaultValue = defaultValue;
+
+      if (type !== 'datepicker') {
+        if ('value' in props) valueProps.value = inputValue;
+        if ('defaultValue' in props) valueProps.defaultValue = defaultValue;
+      }
+
       {
         input = React.createElement(InputTag, Object.assign({
           ref: __reactNode => {
@@ -199,7 +209,7 @@ class F7Input extends React.Component {
           },
           style: inputStyle,
           name: name,
-          type: needsType ? type : undefined,
+          type: needsType ? inputType : undefined,
           placeholder: placeholder,
           id: inputId,
           size: size,
@@ -321,6 +331,12 @@ class F7Input extends React.Component {
       inputEl.removeEventListener('input:empty', self.onInputEmpty, false);
       inputEl.removeEventListener('input:clear', self.onInputClear, false);
     }
+
+    if (self.f7Calendar && self.f7Calendar.destroy) {
+      self.f7Calendar.destroy();
+    }
+
+    delete self.f7Calendar;
   }
 
   componentDidUpdate(prevProps, prevState) {
@@ -332,6 +348,10 @@ class F7Input extends React.Component {
       if (type === 'range' || type === 'toggle') return;
       if (!self.$f7) return;
       self.updateInputOnDidUpdate = true;
+
+      if (self.f7Calendar) {
+        self.f7Calendar.setValue(self.props.value);
+      }
     });
 
     const self = this;
@@ -369,7 +389,8 @@ class F7Input extends React.Component {
         type,
         clearButton,
         value,
-        defaultValue
+        defaultValue,
+        calendarParams
       } = self.props;
       if (type === 'range' || type === 'toggle') return;
       const inputEl = self.refs.inputEl;
@@ -383,6 +404,19 @@ class F7Input extends React.Component {
       if (clearButton) {
         inputEl.addEventListener('input:empty', self.onInputEmpty, false);
         inputEl.addEventListener('input:clear', self.onInputClear, false);
+      }
+
+      if (type === 'datepicker') {
+        self.f7Calendar = f7.calendar.create(Object.assign({
+          inputEl,
+          value,
+          on: {
+            change(calendar, calendarValue) {
+              self.dispatchEvent('calendar:change calendarChange', calendarValue);
+            }
+
+          }
+        }, calendarParams || {}));
       }
 
       f7.input.checkEmptyState(inputEl);
@@ -418,7 +452,7 @@ class F7Input extends React.Component {
 __reactComponentSetProps(F7Input, Object.assign({
   type: String,
   name: String,
-  value: [String, Number, Array],
+  value: [String, Number, Array, Date],
   defaultValue: [String, Number, Array],
   placeholder: String,
   id: [String, Number],
@@ -464,7 +498,8 @@ __reactComponentSetProps(F7Input, Object.assign({
   dropdown: {
     type: [String, Boolean],
     default: 'auto'
-  }
+  },
+  calendarParams: Object
 }, Mixins.colorProps));
 
 F7Input.displayName = 'f7-input';