+++ packages/react/components/page.js
@@ -16,6 +16,7 @@ class F7Page extends React.Component {
         hasSubnavbar: false,
         hasNavbarLarge: false,
         hasNavbarLargeCollapsed: false,
+        hasCardExpandableOpened: false,
         routerPositionClass: '',
         routerForceUnstack: false,
         routerPageRole: null,
@@ -24,7 +25,7 @@ class F7Page extends React.Component {
     })();
 
     (() => {
-      Utils.bindMethods(this, ['onPtrPullStart', 'onPtrPullMove', 'onPtrPullEnd', 'onPtrRefresh', 'onPtrDone', 'onInfinite', 'onPageMounted', 'onPageInit', 'onPageReinit', 'onPageBeforeIn', 'onPageBeforeOut', 'onPageAfterOut', 'onPageAfterIn', 'onPageBeforeRemove', 'onPageStack', 'onPageUnstack', 'onPagePosition', 'onPageRole', 'onPageMasterStack', 'onPageMasterUnstack', 'onPageNavbarLargeCollapsed', 'onPageNavbarLargeExpanded']);
+      Utils.bindMethods(this, ['onPtrPullStart', 'onPtrPullMove', 'onPtrPullEnd', 'onPtrRefresh', 'onPtrDone', 'onInfinite', 'onPageMounted', 'onPageInit', 'onPageReinit', 'onPageBeforeIn', 'onPageBeforeOut', 'onPageAfterOut', 'onPageAfterIn', 'onPageBeforeRemove', 'onPageStack', 'onPageUnstack', 'onPagePosition', 'onPageRole', 'onPageMasterStack', 'onPageMasterUnstack', 'onPageNavbarLargeCollapsed', 'onPageNavbarLargeExpanded', 'onCardOpen', 'onCardClose']);
     })();
   }
 
@@ -194,6 +195,18 @@ class F7Page extends React.Component {
     this.dispatchEvent('page:beforeremove pageBeforeRemove', event, page);
   }
 
+  onCardOpen() {
+    this.setState({
+      hasCardExpandableOpened: true
+    });
+  }
+
+  onCardClose() {
+    this.setState({
+      hasCardExpandableOpened: false
+    });
+  }
+
   render() {
     const self = this;
     const props = self.props;
@@ -286,7 +299,8 @@ class F7Page extends React.Component {
       'page-master': this.state.routerPageRole === 'master',
       'page-master-detail': this.state.routerPageRole === 'detail',
       'page-master-stacked': this.state.routerPageMasterStack === true,
-      'page-with-navbar-large-collapsed': this.state.hasNavbarLargeCollapsed === true
+      'page-with-navbar-large-collapsed': this.state.hasNavbarLargeCollapsed === true,
+      'page-with-card-opened': this.state.hasCardExpandableOpened === true
     }, Mixins.colorClasses(props));
 
     if (!needsPageContent) {
@@ -353,6 +367,8 @@ class F7Page extends React.Component {
     el.removeEventListener('page:masterunstack', self.onPageMasterUnstack);
     el.removeEventListener('page:navbarlargecollapsed', self.onPageNavbarLargeCollapsed);
     el.removeEventListener('page:navbarlargeexpanded', self.onPageNavbarLargeExpanded);
+    el.removeEventListener('card:open', self.onCardOpen);
+    el.removeEventListener('card:close', self.onCardClose);
   }
 
   componentDidMount() {
@@ -391,6 +407,8 @@ class F7Page extends React.Component {
     el.addEventListener('page:masterunstack', self.onPageMasterUnstack);
     el.addEventListener('page:navbarlargecollapsed', self.onPageNavbarLargeCollapsed);
     el.addEventListener('page:navbarlargeexpanded', self.onPageNavbarLargeExpanded);
+    el.addEventListener('card:open', self.onCardOpen);
+    el.addEventListener('card:close', self.onCardClose);
   }
 
   get slots() {