+++ src/phenome/components/input.jsx
@@ -10,7 +10,7 @@ export default {
     // Inputs
     type: String,
     name: String,
-    value: [String, Number, Array],
+    value: [String, Number, Array, Date],
     defaultValue: [String, Number, Array],
     placeholder: String,
     id: [String, Number],
@@ -69,6 +69,9 @@ export default {
       type: [String, Boolean],
       default: 'auto',
     },
+
+    // Datepicker
+    calendarParams: Object,
     ...Mixins.colorProps,
   },
   state() {
@@ -133,12 +136,16 @@ export default {
     let inputEl;
 
     const createInput = (InputTag, children) => {
-      const needsValue = type !== 'file';
+      const needsValue = type !== 'file' && type !== 'datepicker';
       const needsType = InputTag === 'input';
+      let inputType = type;
+      if (inputType === 'datepicker') {
+        inputType = 'text';
+      }
       const inputClassName = Utils.classNames(
         !wrap && className,
         {
-          resizable: type === 'textarea' && resizable,
+          resizable: inputType === 'textarea' && resizable,
           'no-store-data': (noFormStoreData || noStoreData || ignoreStoreData),
           'input-invalid': (errorMessage && errorMessageForce) || self.state.inputInvalid,
           'input-with-value': inputHasValue,
@@ -152,15 +159,17 @@ export default {
         else inputValue = domValue;
       }
       const valueProps = {};
-      if ('value' in props) valueProps.value = inputValue;
-      if ('defaultValue' in props) valueProps.defaultValue = defaultValue;
+      if (type !== 'datepicker') {
+        if ('value' in props) valueProps.value = inputValue;
+        if ('defaultValue' in props) valueProps.defaultValue = defaultValue;
+      }
       if (process.env.COMPILER === 'react') {
         input = (
           <InputTag
             ref="inputEl"
             style={inputStyle}
             name={name}
-            type={needsType ? type : undefined}
+            type={needsType ? inputType : undefined}
             placeholder={placeholder}
             id={inputId}
             size={size}
@@ -204,7 +213,7 @@ export default {
             ref="inputEl"
             style={inputStyle}
             name={name}
-            type={needsType ? type : undefined}
+            type={needsType ? inputType : undefined}
             placeholder={placeholder}
             id={inputId}
             size={size}
@@ -325,6 +334,9 @@ export default {
       if (type === 'range' || type === 'toggle') return;
       if (!self.$f7) return;
       self.updateInputOnDidUpdate = true;
+      if (self.f7Calendar) {
+        self.f7Calendar.setValue(self.props.value);
+      }
     },
   },
   componentDidCreate() {
@@ -333,7 +345,7 @@ export default {
   componentDidMount() {
     const self = this;
     self.$f7ready((f7) => {
-      const { validate, validateOnBlur, resizable, type, clearButton, value, defaultValue } = self.props;
+      const { validate, validateOnBlur, resizable, type, clearButton, value, defaultValue, calendarParams } = self.props;
       if (type === 'range' || type === 'toggle') return;
 
       const inputEl = self.refs.inputEl;
@@ -346,6 +358,19 @@ export default {
       if (clearButton) {
         inputEl.addEventListener('input:empty', self.onInputEmpty, false);
         inputEl.addEventListener('input:clear', self.onInputClear, false);
+      }
+
+      if (type === 'datepicker') {
+        self.f7Calendar = f7.calendar.create({
+          inputEl,
+          value,
+          on: {
+            change(calendar, calendarValue) {
+              self.dispatchEvent('calendar:change calendarChange', calendarValue);
+            },
+          },
+          ...(calendarParams || {}),
+        });
       }
 
       f7.input.checkEmptyState(inputEl);
@@ -401,6 +426,11 @@ export default {
       inputEl.removeEventListener('input:empty', self.onInputEmpty, false);
       inputEl.removeEventListener('input:clear', self.onInputClear, false);
     }
+
+    if (self.f7Calendar && self.f7Calendar.destroy) {
+      self.f7Calendar.destroy();
+    }
+    delete self.f7Calendar;
   },
   methods: {
     domValue() {