+++ src/phenome/components/list-input.jsx
@@ -30,7 +30,7 @@ export default {
       default: 'text',
     },
     name: String,
-    value: [String, Number, Array],
+    value: [String, Number, Array, Date],
     defaultValue: [String, Number, Array],
     readonly: Boolean,
     required: Boolean,
@@ -80,6 +80,10 @@ export default {
     label: [String, Number],
     inlineLabel: Boolean,
     floatingLabel: Boolean,
+
+    // Datepicker
+    calendarParams: Object,
+
     // Colors
     ...Mixins.colorProps,
   },
@@ -156,11 +160,15 @@ export default {
     const isSortable = sortable || self.state.isSortable;
 
     const createInput = (InputTag, children) => {
-      const needsValue = type !== 'file';
+      const needsValue = type !== 'file' && type !== 'datepicker';
       const needsType = InputTag === 'input';
+      let inputType = type;
+      if (inputType === 'datepicker') {
+        inputType = 'text';
+      }
       const inputClassName = Utils.classNames(
         {
-          resizable: type === 'textarea' && resizable,
+          resizable: inputType === 'textarea' && resizable,
           'no-store-data': (noFormStoreData || noStoreData || ignoreStoreData),
           'input-invalid': (errorMessage && errorMessageForce) || inputInvalid,
           'input-with-value': inputHasValue,
@@ -174,15 +182,18 @@ export default {
         else inputValue = domValue;
       }
       const valueProps = {};
-      if ('value' in props) valueProps.value = inputValue;
-      if ('defaultValue' in props) valueProps.defaultValue = defaultValue;
+      if (type !== 'datepicker') {
+        if ('value' in props) valueProps.value = inputValue;
+        if ('defaultValue' in props) valueProps.defaultValue = defaultValue;
+      }
+
       if (process.env.COMPILER === 'react') {
         input = (
           <InputTag
             ref="inputEl"
             style={inputStyle}
             name={name}
-            type={needsType ? type : undefined}
+            type={needsType ? inputType : undefined}
             placeholder={placeholder}
             id={inputId}
             size={size}
@@ -225,7 +236,7 @@ export default {
             ref="inputEl"
             style={inputStyle}
             name={name}
-            type={needsType ? type : undefined}
+            type={needsType ? inputType : undefined}
             placeholder={placeholder}
             id={inputId}
             size={size}
@@ -366,6 +377,9 @@ export default {
       const self = this;
       if (!self.$f7) return;
       self.updateInputOnDidUpdate = true;
+      if (self.f7Calendar) {
+        self.f7Calendar.setValue(self.props.value);
+      }
     },
   },
   componentDidCreate() {
@@ -378,7 +392,7 @@ export default {
     if (!el && !itemContentEl) return;
 
     self.$f7ready((f7) => {
-      const { validate, validateOnBlur, resizable, value, defaultValue, type } = self.props;
+      const { validate, validateOnBlur, resizable, value, defaultValue, type, calendarParams } = self.props;
 
       const inputEl = self.refs.inputEl;
       if (!inputEl) return;
@@ -387,7 +401,18 @@ export default {
       inputEl.addEventListener('textarea:resize', self.onTextareaResize, false);
       inputEl.addEventListener('input:empty', self.onInputEmpty, false);
       inputEl.addEventListener('input:clear', self.onInputClear, false);
-
+      if (type === 'datepicker') {
+        self.f7Calendar = f7.calendar.create({
+          inputEl,
+          value,
+          on: {
+            change(calendar, calendarValue) {
+              self.dispatchEvent('calendar:change calendarChange', calendarValue);
+            },
+          },
+          ...(calendarParams || {}),
+        });
+      }
       if (
         !(validateOnBlur || validateOnBlur === '')
         && (validate || validate === '')
@@ -444,6 +469,10 @@ export default {
     inputEl.removeEventListener('textarea:resize', self.onTextareaResize, false);
     inputEl.removeEventListener('input:empty', self.onInputEmpty, false);
     inputEl.removeEventListener('input:clear', self.onInputClear, false);
+    if (self.f7Calendar && self.f7Calendar.destroy) {
+      self.f7Calendar.destroy();
+    }
+    delete self.f7Calendar;
   },
   methods: {
     domValue() {